move around constants
