+++ spi/common/verifiable-credentials-spi/src/main/java/org/eclipse/edc/iam/verifiablecredentials/spi/model/statuslist2021/StatusList2021Status.java
@@ -26,6 +26,13 @@ import static org.eclipse.edc.iam.verifiablecredentials.spi.VcConstants.STATUSLI
  */
 public class StatusList2021Status {
 
+    public static final String TYPE = "StatusList2021Entry";
+    public static final String STATUS_LIST_CREDENTIAL_LITERAL = "statusListCredential";
+    public static final String STATUS_LIST_CREDENTIAL = STATUSLIST_2021_PREFIX + STATUS_LIST_CREDENTIAL_LITERAL;
+    public static final String STATUS_LIST_INDEX_LITERAL = "statusListIndex";
+    public static final String STATUS_LIST_INDEX = STATUSLIST_2021_PREFIX + STATUS_LIST_INDEX_LITERAL;
+    public static final String STATUS_LIST_PURPOSE_LITERAL = "statusPurpose";
+    public static final String STATUS_LIST_PURPOSE = STATUSLIST_2021_PREFIX + STATUS_LIST_PURPOSE_LITERAL;
     private String statusListPurpose;
     private int statusListIndex;
     private String statusListCredential;
@@ -37,22 +44,22 @@ public class StatusList2021Status {
         var instance = new StatusList2021Status();
         instance.statusListCredential = ofNullable(getId(status))
                 .map(Object::toString)
-                .orElseThrow(() -> new IllegalArgumentException(missingProperty(StatusList2021Credential.STATUS_LIST_CREDENTIAL)));
+                .orElseThrow(() -> new IllegalArgumentException(missingProperty(STATUS_LIST_CREDENTIAL)));
 
-        instance.statusListIndex = ofNullable(status.getProperty(STATUSLIST_2021_PREFIX, StatusList2021Credential.STATUS_LIST_INDEX_LITERAL))
+        instance.statusListIndex = ofNullable(status.getProperty(STATUSLIST_2021_PREFIX, STATUS_LIST_INDEX_LITERAL))
                 .map(Object::toString)
                 .map(Integer::parseInt)
-                .orElseThrow(() -> new IllegalArgumentException(missingProperty(StatusList2021Credential.STATUS_LIST_INDEX)));
+                .orElseThrow(() -> new IllegalArgumentException(missingProperty(STATUS_LIST_INDEX)));
 
-        instance.statusListPurpose = ofNullable(status.getProperty(STATUSLIST_2021_PREFIX, StatusList2021Credential.STATUS_LIST_PURPOSE_LITERAL))
+        instance.statusListPurpose = ofNullable(status.getProperty(STATUSLIST_2021_PREFIX, STATUS_LIST_PURPOSE_LITERAL))
                 .map(Object::toString)
-                .orElseThrow(() -> new IllegalArgumentException(missingProperty(StatusList2021Credential.STATUS_LIST_PURPOSE)));
+                .orElseThrow(() -> new IllegalArgumentException(missingProperty(STATUS_LIST_PURPOSE)));
 
         return instance;
     }
 
     private static Object getId(CredentialStatus status) {
-        var credentialId = status.getProperty(STATUSLIST_2021_PREFIX, StatusList2021Credential.STATUS_LIST_CREDENTIAL_LITERAL);
+        var credentialId = status.getProperty(STATUSLIST_2021_PREFIX, STATUS_LIST_CREDENTIAL_LITERAL);
         if (credentialId instanceof Map<?, ?> map) {
             return map.get("@id");
         }
